+++ packages/forma-36-tokens/src/tokens/colors/colors-purple-deprecated.js
@@ -1,0 +1,11 @@
+const scalePurple = require('./scale-purple');
+
+const colorsPurple = {
+  'color-purple-dark': scalePurple['purple-700'],
+  'color-purple-base': scalePurple['purple-600'],
+  'color-purple-mid': scalePurple['purple-500'],
+  'color-purple-light': scalePurple['purple-300'],
+  'color-purple-lightest': scalePurple['purple-100'],
+};
+
+module.exports = colorsPurple;
