+++ packages/forma-36-tokens/src/tokens/colors/colors-yellow-deprecated.js
@@ -1,0 +1,11 @@
+const scaleYellow = require('./scale-yellow');
+
+const colorsYellow = {
+  'color-yellow-dark': scaleYellow['yellow-800'],
+  'color-yellow-base': scaleYellow['yellow-600'],
+  'color-yellow-mid': scaleYellow['yellow-500'],
+  'color-yellow-light': scaleYellow['yellow-400'],
+  'color-yellow-lightest': scaleYellow['yellow-200'],
+};
+
+module.exports = colorsYellow;
