+++ packages/forma-36-tokens/src/tokens/colors/scale-orange.js
@@ -1,0 +1,13 @@
+const scaleOrange = {
+  'orange-100': '#FFF2E4',
+  'orange-200': '#FDE5C0',
+  'orange-300': '#FDB882',
+  'orange-400': '#F07F23',
+  'orange-500': '#CC4500',
+  'orange-600': '#B12D00',
+  'orange-700': '#892300',
+  'orange-800': '#731A00',
+  'orange-900': '#631C00',
+};
+
+module.exports = scaleOrange;
