+++ packages/forma-36-tokens/src/tokens/colors/scale-purple.js
@@ -1,0 +1,13 @@
+const scalePurple = {
+  'purple-100': '#F7F2FF',
+  'purple-200': '#EDE3FF',
+  'purple-300': '#D1BBFF',
+  'purple-400': '#AE89FF',
+  'purple-500': '#8553E7',
+  'purple-600': '#6C3ECF',
+  'purple-700': '#5127B5',
+  'purple-800': '#3E16A4',
+  'purple-900': '#340E9C',
+};
+
+module.exports = scalePurple;
