+++ packages/forma-36-tokens/src/tokens/colors/scale-yellow.js
@@ -1,0 +1,13 @@
+const scaleYellow = {
+  'yellow-100': '#FEF9DF',
+  'yellow-200': '#FFF6CC',
+  'yellow-300': '#FFE993',
+  'yellow-400': '#FFD960',
+  'yellow-500': '#FFC835',
+  'yellow-600': '#EAAF09',
+  'yellow-700': '#B78300',
+  'yellow-800': '#956300',
+  'yellow-900': '#7F5200',
+};
+
+module.exports = scaleYellow;
